+++ src/test.rs
@@ -24,6 +24,7 @@ mod documentation_examples;
 mod happy_eyeballs;
 mod index_management;
 mod lambda_examples;
+mod plain_auth;
 pub(crate) mod spec;
 mod timeseries;
 pub(crate) mod util;
